auth/pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = "auth service"
authors = ["Dmitry Nikolaev <etsu.nikolaev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.76.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
python-multipart = "^0.0.5"
SQLAlchemy = {extras = ["asyncio"], version = "^1.4.36"}
alembic = "^1.7.7"
uvloop = "^0.16.0"
loguru = "^0.6.0"
asyncpg = "^0.25.0"
pydantic = {extras = ["email"], version = "^1.9.0"}
Jinja2 = "^3.1.2"
fastapi-utils = "^0.2.1"
aiokafka = "^0.7.2"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py310']
exclude = '''
/(
    \.venv
)/
'''

[tool.isort]
profile = "black"
skip = ".venv,"
